+++ ruby_llm_eval/assets/tasks/019_ar_associations/solution_ref.rb
@@ -1,0 +1,7 @@
+class Author < ActiveRecord::Base
+  has_many :books
+end
+
+class Book < ActiveRecord::Base
+  belongs_to :author
+end